onnxruntime_moiblenet/requirements.txt
opencv-python
numpy
onnxruntime
torch
pillow
picamera2